move logger to config/. start work on cors config
