+++ src/server.js
@@ -1,8 +1,10 @@
 import express from 'express';
+import cors from './config/cors.js';
 import routes from './routes/index.js';
 
 const app = express();
 
+app.use(cors);
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
